feat: Update Vue package
