+++ packages/vue/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sentry/vue",
-  "version": "5.24.2",
+  "version": "5.27.3",
   "description": "Offical Sentry SDK for Vue.js",
   "repository": "git://github.com/getsentry/sentry-javascript.git",
   "homepage": "https://github.com/getsentry/sentry-javascript/tree/master/packages/react",
@@ -16,18 +16,18 @@
     "access": "public"
   },
   "dependencies": {
-    "@sentry/browser": "5.24.2",
-    "@sentry/core": "5.24.2",
-    "@sentry/minimal": "5.24.2",
-    "@sentry/types": "5.24.2",
-    "@sentry/utils": "5.24.2",
+    "@sentry/browser": "5.27.3",
+    "@sentry/core": "5.27.3",
+    "@sentry/minimal": "5.27.3",
+    "@sentry/types": "5.27.3",
+    "@sentry/utils": "5.27.3",
     "tslib": "^1.9.3"
   },
   "peerDependencies": {
     "vue": "2.x"
   },
   "devDependencies": {
-    "@sentry-internal/eslint-config-sdk": "5.24.2",
+    "@sentry-internal/eslint-config-sdk": "5.27.3",
     "eslint": "7.6.0",
     "jest": "^24.7.1",
     "jsdom": "^16.2.2",
